config: change login url to our frontend one
Since SSO accounts also need confirmation, after signing up with
one of those, we got redirected to the default Invenio-Accounts
login page that doesn't exist (404).

This wokflow is far from ideal. We should upgrade Invenio-Accounts
to the latest version that support REST apps, and revamp the whole
workflow. As we also need to change to the new CERN SSO, that can
be a good opportunity to tackle this problem.