snakemake: add complexity estimator class
The logic of this estimator class is a bit different than the other workflow engines
as it relies on the `dependencies` between jobs that the Snakemake DAG class calculates
(logic implemented in `reana-commons`). From that information, we extract the maximum
number of jobs that can run in parallel and their assigned kubernetes memory limit to
estimate the complexity tuple with the maximum amount of memory to allocate.

closes #392