rest: check for valid token during authentication
When authenticating a request, the `signin_required` decorator is used
to check whether the user is logged in through Flask or with a REANA
token.

This commit changes the decorator to also check whether the user has a
valid token before performing some of the actions (e.g. workflow
creation).

Closes #516