gitlab: use `url_for` to build API URLs
Use `url_for` to build API URLs. Before, URLs were built in different
ways when creating and retrieving GitLab's webhooks, which sometimes
resulted in having different URLs for the same `create_workflow`
endpoint. This caused already created webhooks not to be recognized by
REANA, as a wrong URL was used to identify them.

Closes #593