global: update to apispec v3
Update to apispec v3, as the version used before is not compatible with
PyYAML v6. This commit also removes the marshmallow plugin for apispec,
as the generated OpenAPI spec is different and causes issues with
reana-client. This is due to the fact that the OpenAPI specifications of
marshmallow schemas are now put inside the global `definitions` object,
instead of being inlined in the endpoint's specification as before.

Closes reanahub/reana-commons#399