feat(config): make number of threads configurable for Dask (#719)
Following the integration of Dask into REANA and discussions with
several REANA users, we identified the need to make the number of
threads configurable.

This commit adds the number of threads configuration to the `/api/info`
endpoint. It also validates that the requested number of threads for a
single Dask worker does not exceed the defined upper limit.

Closes reanahub/reana#874